+++ Example_Card_Calypso/src/main/java/org/eclipse/keyple/card/calypso/example/UseCase4_CardAuthentication/Main_CardAuthentication_Stub.java
@@ -1,0 +1,178 @@
+/* **************************************************************************************
+ * Copyright (c) 2018 Calypso Networks Association https://calypsonet.org/
+ *
+ * See the NOTICE file(s) distributed with this work for additional information
+ * regarding copyright ownership.
+ *
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License 2.0 which is available at http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ************************************************************************************** */
+package org.eclipse.keyple.card.calypso.example.UseCase4_CardAuthentication;
+
+import static org.eclipse.keyple.card.calypso.example.common.ConfigurationUtil.setupCardResourceService;
+
+import org.calypsonet.terminal.calypso.WriteAccessLevel;
+import org.calypsonet.terminal.calypso.card.CalypsoCard;
+import org.calypsonet.terminal.calypso.sam.CalypsoSam;
+import org.calypsonet.terminal.calypso.transaction.CardSecuritySetting;
+import org.calypsonet.terminal.calypso.transaction.CardTransactionManager;
+import org.calypsonet.terminal.reader.selection.CardSelectionManager;
+import org.calypsonet.terminal.reader.selection.CardSelectionResult;
+import org.eclipse.keyple.card.calypso.CalypsoExtensionService;
+import org.eclipse.keyple.card.calypso.example.common.CalypsoCardStubFactory;
+import org.eclipse.keyple.card.calypso.example.common.CalypsoConstants;
+import org.eclipse.keyple.card.calypso.example.common.CalypsoSamStubFactory;
+import org.eclipse.keyple.core.service.Plugin;
+import org.eclipse.keyple.core.service.Reader;
+import org.eclipse.keyple.core.service.SmartCardService;
+import org.eclipse.keyple.core.service.SmartCardServiceProvider;
+import org.eclipse.keyple.core.service.resource.CardResource;
+import org.eclipse.keyple.core.service.resource.CardResourceServiceProvider;
+import org.eclipse.keyple.core.util.ByteArrayUtil;
+import org.eclipse.keyple.plugin.stub.StubPluginFactoryBuilder;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ *
+ *
+ * <h1>Use Case Calypso 4 – Calypso Card authentication (Stub)</h1>
+ *
+ * <p>We demonstrate here the authentication of a Calypso card using a Secure Session in which a
+ * file from the card is read. The read is certified by verifying the signature of the card by a
+ * Calypso SAM.
+ *
+ * <p>Two readers are required for this example: a contactless reader for the Calypso Card, a
+ * contact reader for the Calypso SAM.
+ *
+ * <h2>Scenario:</h2>
+ *
+ * <ul>
+ *   <li>Sets up the card resource service to provide a Calypso SAM (C1).
+ *   <li>Checks if an ISO 14443-4 card is in the reader, enables the card selection manager.
+ *   <li>Attempts to select the specified card (here a Calypso card characterized by its AID) with
+ *       an AID-based application selection scenario.
+ *   <li>Creates a {@link CardTransactionManager} using {@link CardSecuritySetting} referencing the
+ *       SAM profile defined in the card resource service.
+ *   <li>Read a file record in Secure Session.
+ * </ul>
+ *
+ * All results are logged with slf4j.
+ *
+ * <p>Any unexpected behavior will result in runtime exceptions.
+ *
+ * @since 2.0.0
+ */
+public class Main_CardAuthentication_Stub {
+  private static final Logger logger = LoggerFactory.getLogger(Main_CardAuthentication_Stub.class);
+
+  public static void main(String[] args) {
+    final String CARD_READER_NAME = "Stub card reader";
+    final String SAM_READER_NAME = "Stub SAM reader";
+    ;
+
+    // Get the instance of the SmartCardService (singleton pattern)
+    SmartCardService smartCardService = SmartCardServiceProvider.getService();
+
+    // Register the StubPlugin with the SmartCardService, plug a Calypso card stub
+    // get the corresponding generic plugin in return.
+    Plugin plugin =
+        smartCardService.registerPlugin(
+            StubPluginFactoryBuilder.builder()
+                .withStubReader(CARD_READER_NAME, true, CalypsoCardStubFactory.getStubSmartCard())
+                .withStubReader(SAM_READER_NAME, false, CalypsoSamStubFactory.getStubSmartCard())
+                .build());
+
+    // Get the Calypso card extension service
+    CalypsoExtensionService cardExtension = CalypsoExtensionService.getInstance();
+
+    // Verify that the extension's API level is consistent with the current service.
+    smartCardService.checkCardExtension(cardExtension);
+
+    // Get and setup the card reader
+    Reader cardReader = plugin.getReader(CARD_READER_NAME);
+
+    // Configure the card resource service to provide an adequate SAM for future secure operations.
+    setupCardResourceService(plugin, SAM_READER_NAME, CalypsoConstants.SAM_PROFILE_NAME);
+
+    logger.info(
+        "=============== UseCase Calypso #4: Calypso card authentication ==================");
+
+    // Check if a card is present in the reader
+    if (!cardReader.isCardPresent()) {
+      throw new IllegalStateException("No card is present in the reader.");
+    }
+
+    logger.info("= #### Select application with AID = '{}'.", CalypsoConstants.AID);
+
+    // Get the core card selection manager.
+    CardSelectionManager cardSelectionManager = smartCardService.createCardSelectionManager();
+
+    // Create a card selection using the Calypso card extension.
+    // Prepare the selection by adding the created Calypso card selection to the card selection
+    // scenario.
+    cardSelectionManager.prepareSelection(
+        cardExtension
+            .createCardSelection()
+            .acceptInvalidatedCard()
+            .filterByDfName(CalypsoConstants.AID));
+
+    // Actual card communication: run the selection scenario.
+    CardSelectionResult selectionResult =
+        cardSelectionManager.processCardSelectionScenario(cardReader);
+
+    // Check the selection result.
+    if (selectionResult.getActiveSmartCard() == null) {
+      throw new IllegalStateException(
+          "The selection of the application " + CalypsoConstants.AID + " failed.");
+    }
+
+    // Get the SmartCard resulting of the selection.
+    CalypsoCard calypsoCard = (CalypsoCard) selectionResult.getActiveSmartCard();
+
+    logger.info("= SmartCard = {}", calypsoCard);
+
+    logger.info(
+        "Calypso Serial Number = {}",
+        ByteArrayUtil.toHex(calypsoCard.getApplicationSerialNumber()));
+
+    // Create security settings that reference the same SAM profile requested from the card resource
+    // service.
+    CardResource samResource =
+        CardResourceServiceProvider.getService().getCardResource(CalypsoConstants.SAM_PROFILE_NAME);
+    CardSecuritySetting cardSecuritySetting =
+        CalypsoExtensionService.getInstance()
+            .createCardSecuritySetting()
+            .setSamResource(samResource.getReader(), (CalypsoSam) samResource.getSmartCard());
+
+    try {
+      // Performs file reads using the card transaction manager in non-secure mode.
+      cardExtension
+          .createCardTransaction(cardReader, calypsoCard, cardSecuritySetting)
+          .prepareReadRecordFile(
+              CalypsoConstants.SFI_ENVIRONMENT_AND_HOLDER, CalypsoConstants.RECORD_NUMBER_1)
+          .processOpening(WriteAccessLevel.DEBIT)
+          .prepareReleaseCardChannel()
+          .processClosing();
+    } finally {
+      try {
+        CardResourceServiceProvider.getService().releaseCardResource(samResource);
+      } catch (RuntimeException e) {
+        logger.error("Error during the card resource release: {}", e.getMessage(), e);
+      }
+    }
+
+    logger.info(
+        "The Secure Session ended successfully, the card is authenticated and the data read are certified.");
+    logger.info(
+        "File {}h, rec 1: FILE_CONTENT = {}",
+        String.format("%02X", CalypsoConstants.SFI_ENVIRONMENT_AND_HOLDER),
+        calypsoCard.getFileBySfi(CalypsoConstants.SFI_ENVIRONMENT_AND_HOLDER));
+
+    logger.info("= #### End of the Calypso card processing.");
+
+    System.exit(0);
+  }
+}